--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ect"
-version = "1.1.2"
+version = "1.2.0"
 authors = [
   { name="Liz Munch", email="muncheli@msu.edu" },
 ]