--- a/src/ect/ect.py
+++ b/src/ect/ect.py
@@ -113,67 +113,11 @@ class ECT:
 
         return ECTResult(ect_matrix, directions, self.thresholds)
 
-    def _build_incidence_csr(self, graph) -> tuple:
-        """
-        Build column sparse representation of the cell-to-vertex incidence excluding 0-cells. Format is (cell_vertex_pointers, cell_vertex_indices_flat, cell_euler_signs, n_vertices).
-        Example: takes the complex [(1,3),(2,4),(1,2,3)] and returns [(0,2,4,7),(1,3,2,4,1,2,3),(-1,-1,1),4]
-
-        """
-        n_vertices = len(graph.node_list)
-
-        cells_by_dimension = {}
-
-        if hasattr(graph, "edge_indices") and graph.edge_indices is not None:
-            edge_indices_array = np.asarray(graph.edge_indices)
-            if edge_indices_array.size:
-                cells_by_dimension[1] = [
-                    tuple(map(int, row)) for row in edge_indices_array
-                ]
-
-        if hasattr(graph, "cells") and graph.cells:
-            for dim, cells_of_dim in graph.cells.items():
-                if dim == 0:
-                    continue
-                if dim == 1 and 1 in cells_by_dimension:
-                    continue
-                if isinstance(cells_of_dim, np.ndarray):
-                    cell_list = [tuple(map(int, row)) for row in cells_of_dim]
-                else:
-                    cell_list = [tuple(map(int, c)) for c in cells_of_dim]
-                if len(cell_list) > 0:
-                    cells_by_dimension[dim] = cell_list
-
-        dimensions = sorted(cells_by_dimension.keys())
-        n_cells = sum(len(cells_by_dimension[d]) for d in dimensions)
-
-        cell_vertex_pointers = np.empty(n_cells + 1, dtype=np.int64)
-        cell_euler_signs = np.empty(n_cells, dtype=np.int32)
-        cell_vertex_indices_flat = []
-
-        cell_vertex_pointers[0] = 0
-        cell_index = 0
-        for dim in dimensions:
-            cells_in_dim = cells_by_dimension[dim]
-            euler_sign = 1 if (dim % 2 == 0) else -1
-            for cell_vertices in cells_in_dim:
-                cell_vertex_indices_flat.extend(cell_vertices)
-                cell_euler_signs[cell_index] = euler_sign
-                cell_index += 1
-                cell_vertex_pointers[cell_index] = len(cell_vertex_indices_flat)
-
-        cell_vertex_indices_flat = np.asarray(cell_vertex_indices_flat, dtype=np.int32)
-        return (
-            cell_vertex_pointers,
-            cell_vertex_indices_flat,
-            cell_euler_signs,
-            n_vertices,
-        )
-
     def _compute_ect(
         self, graph, directions, thresholds: np.ndarray, dtype=np.int32
     ) -> np.ndarray:
         cell_vertex_pointers, cell_vertex_indices_flat, cell_euler_signs, N = (
-            self._build_incidence_csr(graph)
+            graph._build_incidence_csr()
         )
         thresholds = np.asarray(thresholds, dtype=np.float64)